add query to support top 10 urls

+++ src/main/java/dev/konstantinou/urlshortener/dtos/TopTenDTO.java
@@ -2,5 +2,5 @@ package dev.konstantinou.urlshortener.dtos;
 
 import dev.konstantinou.urlshortener.entities.Url;
 
-public record TopTenDTO (Url url, int visits){
+public record TopTenDTO (Url url, Long visits){
 }
